Have bike-sim.py support gpx or tcx files
Refactor common code out of gpx.py and tcx.py into Path.py and the
gpx/tcx specific code is now in gpx_track_list.py and tcx_track_list.py.

So a creating a Path you pass in a simple List[tp.TrackPoint] which could
be from a gpx or tcx file or some other file in the future, such as a
'.fit' file.

Note: the plotted output is not very smooth so work to do :)